Minor Updates
Fixed a bug causing the drop down not to display correctly when selecting to add a new plant.

Added a toggle state to determine if the list of plants should be updated. this is used when a new plant is added to update the list of plants from the database without needing page refresh