Refactor/ New function
deleteGardenGroup  no longer uses local storage. Deletes Group associated with the users/garden from the Database

createGardenPlant: Adds a new plant to a users garden in the Database.

getPlantData: Updated endpoints, Gets the details of a specified plant from the Database

getPlantDescription: Gets the description of the specified plant from the Database